requirements.txt
torch
torchvision
matplotlib
numpy
scipy
pillow
urllib3
scikit-image
pyyaml